DEV: Increase tests' stability by testing tooltips separately (#23516)
We noticed some of these tests were flaky, sometimes they fail on assertions 
related to testing tooltips. Tooltips are generally hard to test, and it's not 
necessary to test tooltips in every test case. This PR isolates tooltip testing 
in a dedicated test case.

Note we already did the same thing for another spec in a9dfda2 and that 
seems to worked well.