DEV: Fix assertion in embedding test (#24694)
The modified test used to be the same as the test above. The bad test
was introduced in commit 77d4c4d8dc00491b99aa15dfd78eb22a47ffe663,
during a refactoring.

This was not a serious problem because the same behavior was still
tested partially by the other tests below.