FIX: More header dropdown admin sidebar fixes (#27241)
Followup 73c6bb25937a5b6a4872f30126597e76ec6906e6

The admin sidebar was also disappearing on another
child admin route (in this case the docker_manager
plugin update page). Instead of relying on the route
name which is flaky, we can set a boolean when the
sidebar is forced in the root admin route, then
turn it off when leaving admin.